# test/crimson/seastore/test_omap_manager: redesign test replay