@cyx1231st 2025-07-16
https://pulpito.ceph.com/yingxin-2025-07-15_15:39:56-crimson-rados-wip-yingxin-testing-2025-07-15-1743-distro-crimson-debug-smithi/ 2 slow-ops.

All look good.